requirements.txt
# Core dependencies
asyncio>=3.4.3
aiohttp>=3.8.0
aiohttp-cors>=0.7.0
pydantic>=2.0.0
PyYAML>=6.0

# Kubernetes client
kubernetes>=24.2.0

# Google Cloud / AI
google-generativeai>=0.3.0
google-cloud-monitoring>=2.11.0

# Data processing
numpy>=1.21.0
pandas>=1.5.0

# Utilities
python-dateutil>=2.8.0
dataclasses-json>=0.6.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# Optional dependencies for enhanced features
prometheus-client>=0.15.0  # For metrics export
psutil>=5.9.0  # For system monitoring